+++ complex/library/includes/classes/class-admin.php
@@ -6,7 +6,7 @@
  * @subpackage  Admin
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.0.15
  */
 
 
@@ -102,13 +102,14 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 		 * Admin assets
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.0.15
 		 */
 		public static function assets() {
 
 			// Helper variables
 
-				$version = esc_attr( trim( wp_get_theme()->get( 'Version' ) ) );
+				$theme   = {%= prefix_class %}_Theme_Framework::get_theme_slug();
+				$version = wp_get_theme( $theme )->get( 'Version' );
 
 
 			// Register
@@ -127,7 +128,7 @@ final class {%= prefix_class %}_Theme_Framework_Admin {
 					foreach ( $register_styles as $handle => $atts ) {
 						$src   = ( isset( $atts['src'] )   ) ? ( $atts['src']   ) : ( $atts[0] );
 						$deps  = ( isset( $atts['deps'] )  ) ? ( $atts['deps']  ) : ( false    );
-						$ver   = ( isset( $atts['ver'] )   ) ? ( $atts['ver']   ) : ( $version );
+						$ver   = ( isset( $atts['ver'] )   ) ? ( $atts['ver']   ) : ( esc_attr( $version ) );
 						$media = ( isset( $atts['media'] ) ) ? ( $atts['media'] ) : ( 'screen' );
 
 						wp_register_style( $handle, $src, $deps, $ver, $media );