+++ complex/library/includes/classes/class-updater.php
@@ -9,7 +9,7 @@
  * @subpackage  Updater
  *
  * @since    1.0
- * @version  1.0
+ * @version  1.0.15
  */
 
 
@@ -106,7 +106,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 		 * Admin menu link
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.0.15
 		 */
 		public static function menu() {
 
@@ -126,16 +126,18 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 					$xml = self::get_remote_xml_data( {%= prefix_constant %}_UPDATE_NOTIFIER_CACHE_INTERVAL );
 
+					$theme = {%= prefix_class %}_Theme_Framework::get_theme_slug();
+
 					if (
 							isset( $xml->latest )
-							&& version_compare( $xml->latest, wp_get_theme()->get( 'Version' ), '>' )
+							&& version_compare( $xml->latest, wp_get_theme( $theme )->get( 'Version' ), '>' )
 						) {
 
 						add_theme_page(
 							// page_title
 							sprintf(
 								esc_html_x( '%s Theme Updates', '%s stands for the theme name. Just copy it.', '{%= text_domain %}' ),
-								wp_get_theme()->get( 'Name' )
+								wp_get_theme( $theme )->get( 'Name' )
 							),
 							// menu_title
 							esc_html_x( 'Theme Updates', 'Admin menu title.', '{%= text_domain %}' ) . ' <span class="update-plugins count-1"><span class="update-count">1</span></span>',
@@ -159,7 +161,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 		 * Toolbar link
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.0.15
 		 */
 		public static function toolbar() {
 
@@ -182,9 +184,11 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 					$xml = self::get_remote_xml_data( {%= prefix_constant %}_UPDATE_NOTIFIER_CACHE_INTERVAL );
 
+					$theme = {%= prefix_class %}_Theme_Framework::get_theme_slug();
+
 					if (
 							isset( $xml->latest )
-							&& version_compare( $xml->latest, wp_get_theme()->get( 'Version' ), '>' )
+							&& version_compare( $xml->latest, wp_get_theme( $theme )->get( 'Version' ), '>' )
 						) {
 
 						if (
@@ -198,7 +202,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 						$wp_admin_bar->add_menu( array(
 								'id'    => 'update_notifier',
-								'title' => sprintf( esc_html_x( '%s update', 'Admin bar notification link. %s: theme name.', '{%= text_domain %}' ), wp_get_theme()->get( 'Name' ) ) . ' <span id="ab-updates">1</span>',
+								'title' => sprintf( esc_html_x( '%s update', 'Admin bar notification link. %s: theme name.', '{%= text_domain %}' ), wp_get_theme( $theme )->get( 'Name' ) ) . ' <span id="ab-updates">1</span>',
 								'href'  => esc_url( $admin_url )
 							) );
 
@@ -220,7 +224,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 		 * Notifier page renderer
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.0.15
 		 */
 		public static function page() {
 
@@ -244,6 +248,8 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 				$xml = self::get_remote_xml_data( {%= prefix_constant %}_UPDATE_NOTIFIER_CACHE_INTERVAL );
 
+				$theme = {%= prefix_class %}_Theme_Framework::get_theme_slug();
+
 
 			// Processing
 
@@ -255,7 +261,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 				<div class="wrap update-notifier">
 
 					<div id="icon-tools" class="icon32"></div>
-					<h2><strong><?php echo wp_get_theme()->get( 'Name' ); ?></strong> Theme Updates</h2>
+					<h2><strong><?php echo wp_get_theme( $theme )->get( 'Name' ); ?></strong> Theme Updates</h2>
 
 					<br />
 
@@ -267,7 +273,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 							echo '<strong>' . trim( $xml->message ) . '</strong><br />';
 						}
 
-						echo 'You have version ' . wp_get_theme()->get( 'Version' ) . ' installed. <strong>Update to version ' . trim( $xml->latest ) . ' now.</strong>';
+						echo 'You have version ' . wp_get_theme( $theme )->get( 'Version' ) . ' installed. <strong>Update to version ' . trim( $xml->latest ) . ' now.</strong>';
 
 						?></p>
 
@@ -395,7 +401,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 		 * Uses the cached version if available, inside the time interval defined.
 		 *
 		 * @since    1.0
-		 * @version  1.0
+		 * @version  1.0.15
 		 *
 		 * @param  int $interval
 		 */
@@ -419,6 +425,8 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 			// Helper variables
 
+				$theme = {%= prefix_class %}_Theme_Framework::get_theme_slug();
+
 				$db_cache_field              = '{%= prefix_var %}_notifier_cache_' . {%= prefix_constant %}_THEME_SLUG;
 				$db_cache_field_last_updated = '{%= prefix_var %}_notifier_cache_' . {%= prefix_constant %}_THEME_SLUG . '_last_updated';
 				$last                        = get_transient( $db_cache_field_last_updated );
@@ -432,7 +440,7 @@ final class {%= prefix_class %}_Theme_Framework_Updater {
 
 						// Cache doesn't exist, or is old, so refresh it
 
-							$response = wp_remote_get( esc_url( trailingslashit( wp_get_theme()->get( 'AuthorURI' ) ) . 'updates/' . {%= prefix_constant %}_THEME_SLUG . '/' . {%= prefix_constant %}_THEME_SLUG . '-version.xml' ) );
+							$response = wp_remote_get( esc_url( trailingslashit( wp_get_theme( $theme )->get( 'AuthorURI' ) ) . 'updates/' . {%= prefix_constant %}_THEME_SLUG . '/' . {%= prefix_constant %}_THEME_SLUG . '-version.xml' ) );
 
 							if ( is_wp_error( $response ) ) {